refactor(cert): extract legacy migration helpers
